+++ src/btech/combat/mech_hitloc_standard.c
@@ -26,7 +26,7 @@ int mech_hit_location(Mech *mech, int hitGroup, int *iscritical, int *isrear) {
    * config'd */
   /* The advanced FASA table already checks critical immunity; the others do
    * not, so dispatch to the crit-proof table after that check. */
-  switch (mech_class(mech)) {
+  switch ((int)mech_class(mech)) {
   case CLASS_VTOL:
     if (btech_context_uses_advanced_vtol_criticals(context))
       return mech_advanced_vehicle_hit_location(mech, hitGroup, iscritical,
@@ -59,7 +59,7 @@ int mech_hit_location(Mech *mech, int hitGroup, int *iscritical, int *isrear) {
       btech_random_range(context, 1, 100) >= 42)
     return TURRET;
   btech_context_hit_roll_record(context, roll);
-  switch (mech_class(mech)) {
+  switch ((int)mech_class(mech)) {
   case CLASS_BSUIT:
     if ((hitloc = mech_battle_suit_hit_location(mech)) < 0)
       return btech_random_range(context, 0, NUM_BSUIT_MEMBERS - 1);
